+++ app/Http/Controllers/Api/RoleController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Api;
 use App\Trait\ApiResponse;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\StoreRoleRequest;
 use App\Http\Requests\UpdateRoleRequest;
 use App\Repositories\Role\RoleInterface;
@@ -37,6 +38,8 @@ class RoleController extends Controller
      */
     public function index()
     {
+        Gate::authorize('index-role');
+
         $perPage = request('per_page');
         $data = $this->roleRepository->allPaginate($perPage);
         $metadata['count'] = count($data);
@@ -52,6 +55,8 @@ class RoleController extends Controller
      */
     public function store(StoreRoleRequest $request)
     {
+        Gate::authorize('create-role');
+
         $data = $this->roleRepository->store($request);
 
         try {
@@ -79,6 +84,8 @@ class RoleController extends Controller
      */
     public function update(UpdateRoleRequest $request, string $id)
     {
+        Gate::authorize('edit-role');
+
         $data = $this->roleRepository->update($request, $id);
 
         try {
@@ -93,6 +100,8 @@ class RoleController extends Controller
      */
     public function destroy(string $id)
     {
+        Gate::authorize('delete-role');
+
         $data = $this->roleRepository->delete($id);
 
         try {
